Backend/requirements.txt
flask==2.3.3
werkzeug==2.3.7
flask-cors==3.0.10
PyPDF2>=3.0.0
python-docx==0.8.11
google-generativeai>=0.3.0
python-dotenv>=1.0.0
flask-limiter==3.5.0 